# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RagDocument {
  id: number;
  filename: string;
  chunkCount: number;
  createdAt: string;
}

export interface RagChunk {
  id: number;
  text: string;
  similarity: number;
  chunkIndex: number;
  documentId: number;
  documentName: string;
}

export interface RagQueryInput {
  /** @minLength 1 */
  question: string;
}

export interface RagChatInput {
  /** @minLength 1 */
  question: string;
  chunks: RagChunk[];
}

export interface RagRetrieveResult {
  question: string;
  chunks: RagChunk[];
}

export interface RagStats {
  documentCount: number;
  chunkCount: number;
  isModelReady: boolean;
}

export interface DocumentUploadForm {
  file: Blob;
}

export interface RagError {
  error: string;
}