# lib/api-zod/src/generated/types/ragChatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RagChunk } from './ragChunk';

export interface RagChatInput {
  /** @minLength 1 */
  question: string;
  chunks: RagChunk[];
}