# lib/api-zod/src/generated/types/ragChunk.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RagChunk {
  id: number;
  text: string;
  similarity: number;
  chunkIndex: number;
  documentId: number;
  documentName: string;
}