# lib/api-zod/src/generated/types/ragDocument.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RagDocument {
  id: number;
  filename: string;
  chunkCount: number;
  createdAt: Date;
}