# lib/api-zod/src/generated/types/ragRetrieveResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RagChunk } from './ragChunk';

export interface RagRetrieveResult {
  question: string;
  chunks: RagChunk[];
}